Day_04_Forward_Diffusion_Process/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forward-diffusion"
version = "0.1.0"
description = "Forward Diffusion Process Implementation for DDPM"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "Pillow>=9.5.0",
    "imageio>=2.31.0",
    "tqdm>=4.65.0",
    "rich>=13.4.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",
]

[project.scripts]
forward-diffusion = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]